Make `VideoCompositor` a `Bin` capable of framerate conversion.
The old `VideoCompositor` was renamed to `CompositorElement` and a new
`VideoCompositor` was implemented.
