+++ mix.exs
@@ -41,6 +41,7 @@ defmodule Membrane.VideoCompositor.Mixfile do
       {:unifex, "~> 1.0"},
       {:membrane_core, "~> 0.10.0"},
       {:membrane_caps_video_raw, "~> 0.1.0"},
+      {:membrane_framerate_converter_plugin, "~> 0.5.0"},
       {:membrane_raw_video_format, "~> 0.2.0"},
       {:membrane_common_c, "~> 0.13.0"},
       {:nx, "~> 0.2"},
